+++ examples/run_sim.py
@@ -44,3 +44,4 @@ sim.run(verbose=verbose)
 if do_plot:
     print('Plotting...')
     fig = sim.plot(do_save=do_save, do_show=do_show, fig_path=fig_path)
+    sim.plot_result('r_eff')